CTCTRADERS-959: create ModeAtBorderPage and move to folder
